Add files via upload
In This Application, you can get a wide range of quality UI controls for any .NET platform (ASP.NET AJAX, MVC, etc.) This web app will be very helpful for physically disabled users and startups to share the idea and track the details. This project can be used in the fields of Service, Maps, Gauges, Diagrams, Calendars and provides such a great platform for all researchers.